baselines/GraphBAN/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.0
dgl==2.4.0+cu118
dgllife==0.3.2
fair-esm==2.0.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.2.0
future==1.0.0
huggingface-hub==0.26.5
hyperopt==0.2.7
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
lightning-utilities==0.11.9
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==2.2.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
pandas==2.2.2
pillow==10.2.0
prettytable==3.12.0
propcache==0.2.1
psutil==6.1.0
py4j==0.10.9.7
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pyg-lib==0.4.0+pt23cu118
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rdkit==2024.3.6
rdkit-pypi==2022.9.5
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.3.1+cu118
torch-geometric==2.6.1
torch_cluster==1.6.3+pt23cu118
torch_scatter==2.1.2+pt23cu118
torch_sparse==0.6.18+pt23cu118
torch_spline_conv==1.2.2+pt23cu118
torchaudio==2.3.1+cu118
torchmetrics==1.6.0
torchvision==0.18.1+cu118
tqdm==4.67.1
transformers==4.47.0
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
yacs==0.1.8
yarl==1.18.3
